+++ src/index.ts
@@ -3,4 +3,4 @@ export * from './execFile';
 export * from './fork';
 export * from './spawn';
 export * from './operators';
-export * from './util';
+export {spawnEnd} from './util';